added System.exit(0) after update success
Java has this feature that allows a program to run even after it's .class file has been changed, or deleted. That is bad for Samex because after the files are deleted, and downloaded again, from samexmain.github.io, the JVM runs the previous version, this can't be avoided by deleting and creating the object again, that's what I thought would happen after implementing restart, but it doesn't work. So I added System.exit(0) after a successful update so that the client wont be weirded out.